+++ src/OpenTelemetry.AutoInstrumentation/Configuration/EnvironmentConfigurationTracerHelper.cs
@@ -32,6 +32,7 @@ internal static class EnvironmentConfigurationTracerHelper
 #if NETCOREAPP3_1_OR_GREATER
         [TracerInstrumentation.MongoDB] = builder => builder.AddSource("MongoDB.Driver.Core.Extensions.DiagnosticSources"),
         [TracerInstrumentation.MySqlData] = builder => builder.AddSource("OpenTelemetry.Instrumentation.MySqlData"),
+        [TracerInstrumentation.StackExchangeRedis] = builder => builder.AddSource("OpenTelemetry.Instrumentation.StackExchangeRedis"),
 #endif
         [TracerInstrumentation.Npgsql] = builder => builder.AddSource("Npgsql")
     };