+++ src/OpenTelemetry.AutoInstrumentation/Loading/LazyInstrumentationLoader.cs
@@ -53,7 +53,7 @@ internal class LazyInstrumentationLoader : ILifespanManager, IDisposable
         _ = new OnAssemblyLoadInitializer(this, loader);
     }
 
-    void ILifespanManager.Track(object instance)
+    public void Track(object instance)
     {
         _instrumentations.Add(instance);
     }