+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Add MySql.Data traces instrumentation.
 - Add Npgsql traces instrumentation.
+- Add StackExchange.Redis traces instrumentation.
 - Add configuration option `none` to `OTEL_DOTNET_AUTO_TRACES_ENABLED_INSTRUMENTATIONS`
   and `OTEL_DOTNET_AUTO_METRICS_ENABLED_INSTRUMENTATIONS`.